Updated <em> tag logic using _
In markdown, an _ is something special ... One cannot just scan for every _ in a text and emphasize every text between _ (underscores).
In many program languages, _ are used to represent a space in identifiers and functions etc.
Therefore, I've updated the markdown.lua logic to ensure that text between _ is only emphasized when the first _ starts with a space or any other non alpameric character, or the _ starts at the beginning of the text.

examples:

AI\_CARGO\_DISPATCHER gives AI\_CARGO\_DISPATCHER.
AI \_CARGO\_ DISPATCHER gives AI <em>CARGO</em> DISPATCHER.
\_AI\_CARGO\_DISPATCHER gives <em>AI</em>CARGO\_DISPATCHER